pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cevolve"
version = "0.1.0"
description = "Genetic algorithms for autonomous code optimization"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "mlx>=0.30.0",
    "numpy>=2.2.6",
    "pyarrow>=21.0.0",
    "pyyaml>=6.0.0",
    "regex>=2025.7.34",
    "requests>=2.32.0",
    "rich>=13.0.0",
    "rustbpe>=0.1.0",
    "textual>=0.50.0",
    "tiktoken>=0.11.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["evolve"]

[tool.uv]
package = true

[project.scripts]
cevolve = "evolve.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
